feat: mejoras

+++ src/main/java/com/proyectojwt/repository/EmpleadoRepository.java
@@ -6,7 +6,6 @@ import com.proyectojwt.entity.Empleado;
 import java.util.List;
 
 
-public interface EmpleadoRepository extends JpaRepository<Empleado, Integer>{
-	
+public interface EmpleadoRepository extends JpaRepository<Empleado, Long>{
          List<Empleado> findByCiudadCodigo(int ciu);
 }
